+++ docs/docs-components/MarkdownPage.js
@@ -34,7 +34,7 @@ export default function MarkdownPage({ children, meta, pageSourceUrl }: Props): 
   const components = {
     small: (props) => <Text size="100">{props.children}</Text>,
     pre: (props: {|
-      children: {| props: {| className: string[], children: string | null |} |},
+      children: {| props: {| className: Array<string>, children: string | null |} |},
     |}) => (
       <Highlighter classNames={props.children.props.className}>
         {props.children.props.children}